Update GitHub repo creation steps
These instructions made some assumptions that could lead to issues in the next steps. For example, it wasn't clear that the root of the new repo needed to be a specific subfolder from the Dagster U repo. After updating that, I also wanted to be more explicit about the .gitignore in the new repo, since without it, the DuckDB file would likely fail to push due to its size.